Move vm/ec2vm/params.go to vm/ec2params/params.go
